requirements.txt
python-dotenv
telethon
tqdm
pytest
